tutorials/delivering-cloud-monitoring-notifications-to-third-party-services/jira_integration_example/requirements.txt
Flask==1.1.2
pylint==2.5.3
pytest==5.3.2; python_version > "3.0"
pytest==5.3.2; python_version < "3.0"
pytest-mock==3.2.0
gunicorn==20.0.4
google-cloud-secret-manager==1.0.0
google-cloud-monitoring==1.0.0
python-dotenv==0.13.0
jira==2.0.0